requirements.txt
pymongo
python-dotenv
pandas
numpy
scikit-learn
simplejson
flask